Needed to change ints to longs to handle billions. Added unit test.

+++ NumberToWordsLib/Handlers/MillionsHandler.cs
@@ -3,12 +3,12 @@
     // Concrete handler for millions
     public class MillionsHandler : NumberHandler
     {
-        public override string Handle(int number)
+        public override string Handle(long number)
         {
             if (number >= 1000000 && number < 1000000000)
             {
-                int millions = number / 1000000;
-                int remainder = number % 1000000;
+                long millions = number / 1000000;
+                long remainder = number % 1000000;
                 return NumberToWordsBelow1000(millions) + " million" + (remainder > 0 ? " " + Handle(remainder) : "");
             }
             else
